Preserve upstream's original makedumpfile.8
The Makefile modifies makedumpfile.8 in place.  If the .orig.tar.gz had
the already-updated version, this bug wouldn't manifest itself, but in
version 1.3.4, makedumpfile.8 had a header from version 1.3.3.  So, if
we built the package, then ran "debian/rules clean", we wound up with
different source than we started with.

So, we now make a copy of makedumpfile.8 before dh_auto_build, and copy
the original back before dh_auto_clean.