Update packaging for upstream version 1.4.0
This change also drops the makedumpfile-static package.  It now depends
on libebl, which is not available in a static library.  In any case,
Nobody is using it (I added it so Ubuntu could use it in their custom
crash initrd, but they decided to use the dynamic version in a chroot
instead).