[debian/hook] Add postrm and postinst kernel hooks
The postinst hook generates smaller initrd.img files
in order to minimize the risk of OOM in the kexec
booted kernel.

The postrm hook removes the corresponding
initrd.img when the kernel package is removed.

Smaller initrd.img as placed in KDUMP_DIR which defaults
to /var/lib/kdump

The $KDUMP_DIR/initramfs-tools needs to be recreated everytime
in the eventuality that it has been updated since the last
time we ran.

Signed-off-by: Louis Bouchard <louis.bouchard@canonical.com>