[debian/kdump-config] Add symlink management for smaller initrd.img
If the $KDUMP_DIR directory exists, symlinks will be verified
for validity (exists/broken/absent).  Log messages will report
the status of symlinks for initrd.img and vmlinuz.

Unless DRY_RUN=true, the symlinks will be created if absent,
or removed and re-created if they are broken, using the
current kernel version when the 'load' command is used.

Their presence and validity will be verified when 'test' and
'status' commands are issued.

Signed-off-by: Louis Bouchard <louis.bouchard@canonical.com>