[kdump-config] Simplify link creation logic
We no longer discriminate on the type of link failure (broken/absent)
and force link re-creation systematically.

check_symlink() is also enhanced to take the kernel version as an
argument. It also check to see if the link points to the proper kernel
version and will create a new symlink if the version differs.

Signed-off-by: Louis Bouchard <louis.bouchard@ubuntu.com>