[debian] Make kexec-tools dependancy disto specific
On Ubuntu, kexec-tools used to have /etc/default/grub.d/kexec-tools.cfg
that is now in kdump-tools. So a versionned Depends: is needed to force
the upgrade to the latest kexec-tools.

On Debian, there is no such requirement as it never carried the
crashkernel parameter definition so there is no restriction on the
kexec-tools version.

Make sure that the same source package can generate the appropriate
Depends: depending on the distro.