Select appropriate /etc/default/grub.d/kdump-tools.cfg at build time.
In 1.5.9-6, a facility was added to make /etc/default/grub.d/kdump-tools.cfg
a symlink, pointing to an appropriate architecture-specific config, with
ppc64el as the initial user. There are a couple of issues with this:

  - It doesn't work on ppc64el today. Reason being that the postinst
    uses the "arch" command to ID the architecture. That returns the
    kernel arch name (ppc64*le*), not the Debian port name
    (ppc64*el*), so the intended symlink is not created.
  - This method requires that we install all of the arch-specific
    configs on every arch, even though they aren't needed.

We simplify things here by doing the default config file selection at build-
time instead of install-time, while still preserving the behavior of dpkg
only considering the file to be modified if the user manually edited it.

It also moves to a more declarative method of mapping architectures to the
default config by using symlinks instead of maintaining a list in the rules
file.

On the negative side, this does mean kdump-tools is now no longer Arch:all,
since each architecture can now have it's own default config. We restrict it
to building on the same architectures as makedumpfile, since kdump-tools
depends on it.

This also fixes a bug where the /etc/default/grub.d files would not be
installed on any architecture if kdump-tools happened to have been built on
any arch other than i386, amd64, powerpc, ppc64, ppc64el or ia64.