Undo the /etc/default/grub.d/kdump-tools.cfg symlink transition
Remove the symlink if it exists, and hasn't been modified. While it
would be nice to also clean up the symlink on ppc64el systems if the user
had manually fixed it to point at kdump-tools..ppc64el, I'm treating the
symlink with the same respect as any other config file by preserving any
user changes.

On !ppc64el, we tell dpkg to move the kdump-tools.default file to
kdump-tools.cfg, and we cleanup the kdump-tools..ppc64el file. On ppc64el,
we need to mv the kdump-tools..ppc64el instead so that dpkg doesn't think
the user manually modified the file if they haven't, and we cleanup the
kdump-tools.default file.