Use multi-user.target.
Some systemd versions will fail to bring up kdump-tools when the target
is an alias, which is the case for default. Also, default target could
be changed to a recovery target, for example, and it's possible we don't
want to invoke kdump-tools in such a case. Using multi-user is a
sensible default.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@debian.org>