kdump-tools Recommends initramfs-tools-core
Though an initramfs is not needed when using kdump-tools, the user will
have to set it up in a way that it's not going to be required when
loading the kdump kernel. By default, the config script will try to
create an initramfs and require it to be present so load can work.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@debian.org>